filter and search works with paging but the code needs to be refactored
