+++ app/src/main/java/com/movies/MoviesApp.kt
@@ -2,7 +2,7 @@ package com.movies
 
 import android.app.Application
 import com.movies.data.local.module.dataBaseModule
-import com.movies.data.mapper.module.entityMapperModule
+import com.movies.data.local.mapper.module.entityMapperModule
 import com.movies.data.remote.di.dtoMapperModule
 import com.movies.data.remote.di.networkModule
 import com.movies.domain.di.repositoryModule
