+++ app/src/main/java/com/movies/data/remote/dto/MoviesDTO.kt
@@ -1,4 +1,4 @@
-package com.movies.data.remote.model
+package com.movies.data.remote.dto
 
 
 import com.google.gson.annotations.SerializedName
