+++ app/src/main/java/com/movies/domain/usecase/movies/GetMoviesUseCase.kt
@@ -1,6 +1,5 @@
 package com.movies.domain.usecase.movies
 
-import android.util.Log
 import androidx.paging.Pager
 import com.movies.common.network.CategoryType
 import com.movies.domain.model.MovieDomainModel
@@ -12,7 +11,6 @@ class GetMoviesUseCase(
 ) : BaseUseCase<CategoryType, Pager<Int, MovieDomainModel>>() {
 
     override suspend fun invoke(params: CategoryType?): Pager<Int, MovieDomainModel> {
-        Log.d("TAG", "useCase: $params")
         return moviesRepository.fetchMovies(params!!)
     }
