+++ app/src/main/java/com/movies/presentation/details/view_model/DetailsViewModel.kt
@@ -1,7 +1,7 @@
 package com.movies.presentation.details.view_model
 
 import com.movies.common.extensions.viewModelScope
-import com.movies.domain.usecase.movies.UpdateFavouriteStatusMovieUseCase
+import com.movies.domain.usecase.favourites.UpdateFavouriteStatusMovieUseCase
 import com.movies.presentation.base.view_model.BaseViewModel
 import com.movies.presentation.model.mapper.MovieUIToDomainMapper
 import com.movies.presentation.model.movie.MovieUIModel
