+++ app/src/main/java/com/movies/presentation/home/adapter/movie/MoviePagingAdapter.kt
@@ -1,12 +1,14 @@
-package com.movies.presentation.base.adapter.movie_adapter
+package com.movies.presentation.home.adapter.movie
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.movies.common.extensions.loadImage
 import com.movies.databinding.MovieItemBinding
+import com.movies.presentation.base.adapter.movie_adapter.BaseMovieAdapter
+import com.movies.presentation.base.adapter.paging_adapter.BasePagingMovieAdapter
 import com.movies.presentation.model.movie.MovieUIModel
 
-class FavouriteMovieAdapter : BaseMovieAdapter<MovieUIModel, FavouriteMovieAdapter.MoviesViewHolder>() {
+class MoviePagingAdapter : BasePagingMovieAdapter<MovieUIModel, MoviePagingAdapter.MoviesViewHolder>() {
 
     override fun onCreateViewHolder(
         parent: ViewGroup,
@@ -22,7 +24,7 @@ class FavouriteMovieAdapter : BaseMovieAdapter<MovieUIModel, FavouriteMovieAdapt
     }
 
     class MoviesViewHolder(val binding: MovieItemBinding) :
-        BaseMovieViewHolder<MovieUIModel>(binding) {
+        BaseMovieAdapter.BaseMovieViewHolder<MovieUIModel>(binding) {
         override fun onBindMovie(
             item: MovieUIModel,
             onClickCallback: ((MovieUIModel) -> Unit)?,
