+++ app/src/main/java/com/movies/presentation/home/view_model/HomeViewModel.kt
@@ -1,5 +1,6 @@
 package com.movies.presentation.home.view_model
 
+import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import androidx.paging.PagingData
 import androidx.paging.cachedIn
@@ -7,9 +8,9 @@ import androidx.paging.map
 import com.movies.common.extensions.viewModelScope
 import com.movies.common.network.CategoryType
 import com.movies.common.utils.LiveDataDelegate
-import com.movies.domain.usecase.movies.GetFavouriteMoviesUseCase
+import com.movies.domain.usecase.favourites.GetFavouriteMoviesUseCase
 import com.movies.domain.usecase.movies.GetMoviesUseCase
-import com.movies.domain.usecase.movies.UpdateFavouriteStatusMovieUseCase
+import com.movies.domain.usecase.favourites.UpdateFavouriteStatusMovieUseCase
 import com.movies.domain.usecase.search.SearchMoviesUseCase
 import com.movies.presentation.base.view_model.BaseViewModel
 import com.movies.presentation.home.ui.HomeFragmentDirections
@@ -32,13 +33,19 @@ class HomeViewModel(
     val fetchFavouriteMoviesLivedata by LiveDataDelegate<List<MovieUIModel>>()
 
     private val _fetchMoviesStateFlow = MutableStateFlow<PagingData<MovieUIModel>?>(null)
-    val fetchMoviesStateFlow get() = _fetchMoviesStateFlow.asStateFlow()
+
+    //    val fetchMoviesStateFlow get() = _fetchMoviesStateFlow.asStateFlow()
+    val fetchMoviesStateFlow
+        get() = _fetchMoviesStateFlow.asLiveData()
 
     private val _categoryStateFlow = MutableStateFlow(CategoryType.POPULAR)
     val categoryStateFlow = _categoryStateFlow.asStateFlow()
 
     private val _searchStateFlow = MutableStateFlow<PagingData<MovieUIModel>?>(null)
-    val searchStateFlow = _searchStateFlow.asStateFlow()
+
+    //    val searchStateFlow = _searchStateFlow.asStateFlow()
+    val searchStateFlow
+        get() = _searchStateFlow.asLiveData()
 
     init {
         getMovies()
@@ -53,7 +60,6 @@ class HomeViewModel(
                     val mappedData = pagingData.map {
                         moviesUIMapper(it)
                     }
-
                     _fetchMoviesStateFlow.emit(mappedData)
                 }
             loadingLiveData.addValue(false)
