+++ app/src/main/res/layout/fragment_details.xml
@@ -27,8 +27,8 @@
         android:id="@+id/backImageButton"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="@dimen/dimen_16_dp"
-        android:layout_marginTop="@dimen/dimen_8_dp"
+        android:paddingStart="@dimen/dimen_16_dp"
+        android:paddingTop="@dimen/dimen_8_dp"
         android:background="@null"
         android:src="@drawable/ic_back"
         app:layout_constraintEnd_toStartOf="@id/titleTextView"
